third_party/blink/web_tests/fast/table/padding-height-and-override-height-expected.txt

FAIL document.getElementById('cell-contents').offsetHeight should be 78. Was 88.
FAIL document.getElementById('cell-contents').offsetWidth should be 88. Was 98.
PASS successfullyParsed is true

TEST COMPLETE
